\name{ll}
\alias{ll}
\title{ll}
\usage{
  ll(beta, gamma, sigma, alpha, mdzero, maxm, y, X)
}
\arguments{
  \item{beta}{}

  \item{gamma}{}

  \item{sigma}{baseline normal(0, sd = sigma)}

  \item{alpha}{PT parameter}

  \item{mdzero}{}

  \item{maxm}{}

  \item{y}{}

  \item{X}{}
}
\value{
  log likelihood
}
\description{
  Log Likelihood given beta, gamma, sigma, alpha, mdzero
}
\author{
  Minzhao Liu, Mike Daniels
}

