\name{PostDensity}
\alias{PostDensity}
\title{PostDensity}
\usage{
  PostDensity(grid, beta, gamma, sigma, alpha, mdzero,
    maxm, y, X)
}
\arguments{
  \item{grid}{}

  \item{beta}{}

  \item{gamma}{}

  \item{sigma}{}

  \item{alpha}{}

  \item{mdzero}{}

  \item{maxm}{}

  \item{y}{}

  \item{X}{}
}
\value{
  postdensity
}
\description{
  postdensity
}
\author{
  Minzhao Liu, Mike Daniels
}

