\name{PostQuantile}
\alias{PostQuantile}
\title{PostQuantiles}
\usage{
  PostQuantile(beta, gamma, sigma, alpha, mdzero, maxm, y,
    X, quan)
}
\arguments{
  \item{beta}{}

  \item{gamma}{}

  \item{sigma}{}

  \item{alpha}{}

  \item{mdzero}{}

  \item{maxm}{}

  \item{y}{}

  \item{X}{}
}
\value{
  postquantile
}
\description{
  calculate the posterior quantiles
}
\author{
  Minzhao Liu, Mike Daniels
}

